Split create util to workaround kube-rs issue

+++ policy-test/src/lib.rs
@@ -17,9 +17,9 @@ pub enum LinkerdInject {
     Disabled,
 }
 
-pub async fn create<T>(client: &kube::Client, obj: T) -> T
+pub async fn create_global<T>(client: &kube::Client, obj: T) -> T
 where
-    T: kube::Resource,
+    T: kube::Resource<Scope = kube::core::ClusterResourceScope>,
     T: serde::Serialize + serde::de::DeserializeOwned + Clone + std::fmt::Debug,
     T::DynamicType: Default,
 {
@@ -28,6 +28,26 @@ where
         ..Default::default()
     };
     let api = kube::Api::<T>::all(client.clone());
+    tracing::trace!(?obj, "Creating");
+    api.create(&params, &obj)
+        .await
+        .expect("failed to create resource")
+}
+
+pub async fn create<T>(client: &kube::Client, obj: T) -> T
+where
+    T: kube::Resource<Scope = kube::core::NamespaceResourceScope>,
+    T: serde::Serialize + serde::de::DeserializeOwned + Clone + std::fmt::Debug,
+    T::DynamicType: Default,
+{
+    let params = kube::api::PostParams {
+        field_manager: Some("linkerd-policy-test".to_string()),
+        ..Default::default()
+    };
+    let api = obj
+        .namespace()
+        .map(|ns| kube::Api::<T>::namespaced(client.clone(), &*ns))
+        .unwrap_or_else(|| kube::Api::<T>::default_namespaced(client.clone()));
     tracing::trace!(?obj, "Creating");
     api.create(&params, &obj)
         .await
@@ -159,7 +179,7 @@ where
 
     let name = format!("linkerd-policy-test-{}", random_suffix(6));
     tracing::debug!(namespace = %name, "Creating");
-    let ns = create(
+    let ns = create_global(
         &client,
         k8s::Namespace {
             metadata: k8s::ObjectMeta {
@@ -210,10 +230,12 @@ where
         drop(_tracing);
     }
 
-    tracing::debug!(ns = %ns.name_unchecked(), "Deleting");
-    api.delete(&ns.name_unchecked(), &kube::api::DeleteParams::background())
-        .await
-        .expect("failed to delete Namespace");
+    if std::env::var("POLICY_TEST_NO_CLEANUP").is_err() {
+        tracing::debug!(ns = %ns.name_unchecked(), "Deleting");
+        api.delete(&ns.name_unchecked(), &kube::api::DeleteParams::background())
+            .await
+            .expect("failed to delete Namespace");
+    }
     if let Err(err) = res {
         std::panic::resume_unwind(err.into_panic());
     }
